refactor: (WIP) overhaul structure of whole project
I've just kind of been in the zone and wanted to grind out as much as
possible without bothering with git and branches and everything. This is
definitely still in an incomplete state. I've just been working for
a while and I wanted to save stuff here.

I still don't think this is exactly "hexagonal architecture", but there
is certainly more separation between the "business logic" and the
implementation details. Still much to be done before a merge to main.