feat: implement 'prepare image and append to movie'
Here I added two functions in a new 'main' module in core. These
functions perform the two main operations of the app: prepare images and
create a movie from them, and prepare one image and append it to an
existing movie. This commit includes the implementation of using
`ffmpeg` to append an image to a movie.

There is a bug where the newly appended movie comes out correctly, but
its duration is greater than it should be (basically, the video finishes
but the video player thinks the movie is still playing).

One idea is to create a movie made of one frame and concat that onto the
existing movie, instead of trying to append the frame itself to the
movie. It's possible that that is already what we're doing -- I wouldn't
know because I didn't look too far into the `ffmpeg` command.