feat: complete CLI implementation
The CLI now works correctly, but there is an issue with overwriting
a movie (i.e., appending to a movie and not supplying a new output
path). I believe the problem is that FFmpeg is still using the original
while it is writing to the new file, so the output is corrupt and
incorrect. This answer (https://stackoverflow.com/a/28877452/3801865)
corroborates this, so the next commit will have to use a temporary file
to do the overwriting.