refactor: pass font path instead of using env vars
The previous commit added the functionality of .env files so that AWS
and the CLI could both set the value of the path to the font file
separately. This was done in order to avoid passing the font file around
from the highest level down to the "write text on picture" method.

I realized I could use a class variable instead, so this commit undoes
the .env stuff in favor of this class variable. AWS can still use an
environment variable to define this, but now the CLI can receive a path
to the font file, making this much more flexible.