feat: (WIP) create adapter for AWS Lambda function
The new file in this commit is the code for the AWS Lambda function. It
is complete except for the "meat" of the function, which actually does
the image manipulation/movie creation. The rest of it has all been
tested on Lambda/S3 and it works. The next steps are writing the code
which calls our core function, adding an ffmpeg layer to Lambda, and
testing.